requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.26.0
botocore==1.29.165
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
Django==4.2.16
djangorestframework==3.14.0
easyocr==1.7.2
exceptiongroup==1.2.2
fastapi==0.95.0
filelock==3.16.1
fsspec==2024.10.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
language_tool_python==2.8.1
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.9.1
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
prometheus-client==0.15.0
pyclipper==1.3.0.post6
pydantic==1.10.19
pytesseract==0.3.10
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
s3transfer==0.6.2
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.26.1
sympy==1.13.1
textblob==0.18.0.post0
tifffile==2024.8.30
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.18.3
